made some imports
