+++ js/newpost.js
@@ -1,0 +1,5 @@
+function stringLengthValidation(string, maxLength) {
+  return string.length >= maxLength;
+}
+
+stringLengthValidation('string', 140);
